V3.0.0

See merge request ahinoam/trivia-project!10

+++ ClientWPF/ClientWPF/CreateRoom.xaml.cs
@@ -63,7 +63,8 @@ namespace ClientWPF
             };
 
             string json = JsonConvert.SerializeObject(createRoom, Formatting.Indented);
-            CreateRoomResponse createRoomResponse = Communicator.ManageSendAndGetData<CreateRoomResponse>(json, clientStream, Codes.CREATE_ROOM_CODE);
+            CreateRoomResponse createRoomResponse = Communicator.ManageSendAndGetData<CreateRoomResponse>(
+                json, clientStream, (int)Codes.CREATE_ROOM_CODE);
 
             if (createRoomResponse.Status == ERROR_CODE)
             {
@@ -75,20 +76,20 @@ namespace ClientWPF
             { 
                 int roomId = createRoomResponse.RoomId;
 
-            RoomData roomData = new RoomData
-            {
-                RoomId = roomId,
-                NumOfplayers = createRoom.NumOfPlayers,
-                NumOfQuestions = createRoom.NumOfQuestions,
-                RoomName = createRoom.RoomName,
-                TimeForQuestion = createRoom.TimeForQuestion,
-                IsActive = 1
-            };
+                RoomData roomData = new RoomData
+                {
+                    RoomId = roomId,
+                    NumOfplayers = createRoom.NumOfPlayers,
+                    NumOfQuestions = createRoom.NumOfQuestions,
+                    RoomName = createRoom.RoomName,
+                    TimeForQuestion = createRoom.TimeForQuestion,
+                    IsActive = 1
+                };
 
-            MessageBox.Show("about to open room admin.");
-            // Show the room details to the admin.
-            var roomAdmin = new RoomAdmin(roomData, clientStream, true);
-            roomAdmin.Show();
+                MessageBox.Show("about to open room admin.");
+                // Show the room details to the admin.
+                var roomAdmin = new RoomAdmin(roomData, clientStream, true);
+                roomAdmin.Show();
             }
             Close();
         }
